requirements.txt
django==3.1.7
gunicorn==20.1.0
djangorestframework==3.12.4
django-cors-headers==3.7.0
psycopg2-binary==2.9.6
celery==4.4.1
redis==3.4.1
requests==2.31.0
pandas==2.1.1

# Testing
pytest==7.2.0
pytest-django==4.1.0
factory_boy==3.2.0
freezegun==1.1.0
